Pro_Venv/requirements.txt
asgiref==3.2.10
Django==3.0.8
django-cors-headers==3.4.0
djangorestframework==3.11.0
Pillow==7.2.0
psycopg2==2.8.5
pytz==2020.1
reportlab==3.5.44
sqlparse==0.3.1
